requirements.txt
httpx[http2]
python-dotenv>=1.1.1
tqdm>=4.67.1
rich>=14.1.0
